+++ client/include/webcface/internal/data_store2.h
@@ -1,8 +1,7 @@
 #pragma once
+#include <deque>
 #include <mutex>
-#include <unordered_map>
 #include <optional>
-#include <functional>
 #include "webcface/field.h"
 #include "webcface/common/val_adaptor.h"
 #include "webcface/internal/func_internal.h"
@@ -11,7 +10,6 @@
 #include "webcface/component_canvas3d.h"
 #include "webcface/component_view.h"
 #include "webcface/robot_link.h"
-#include "webcface/common/internal/message.h"
 #include "webcface/internal/component_internal.h"
 #include "webcface/internal/robot_link_internal.h"
 #include "webcface/log.h"
