+++ client/src/client_msg_recv.cc
@@ -1,12 +1,18 @@
-#include "webcface/image.h"
+#include "webcface/common/internal/message/pack.h"
+#include "webcface/common/internal/message/func.h"
+#include "webcface/common/internal/message/log.h"
+#include "webcface/common/internal/message/sync.h"
+#include "webcface/common/internal/message/text.h"
+#include "webcface/common/internal/message/value.h"
 #include "webcface/log.h"
 #include "webcface/member.h"
-#include "webcface/value.h"
 #include "webcface/view.h"
 #include "webcface/func.h"
+#include "webcface/image.h"
+#include "webcface/value.h"
+#include "webcface/text.h"
 #include "webcface/canvas3d.h"
 #include "webcface/canvas2d.h"
-#include "webcface/common/internal/message.h"
 #include "webcface/internal/client_internal.h"
 #include "webcface/internal/robot_link_internal.h"
