+++ client/src/client_msg_sync.cc
@@ -1,8 +1,12 @@
+#include "webcface/common/internal/message/pack.h"
+#include "webcface/common/internal/message/func.h"
+#include "webcface/common/internal/message/log.h"
 #include "webcface/client.h"
+#include "webcface/common/internal/message/sync.h"
+#include "webcface/common/internal/message/text.h"
+#include "webcface/common/internal/message/value.h"
 #include "webcface/internal/logger.h"
-#include "webcface/common/internal/message.h"
 #include "webcface/internal/client_internal.h"
-#include "webcface/internal/robot_link_internal.h"
 
 WEBCFACE_NS_BEGIN
