+++ client/src/func_info.cc
@@ -1,5 +1,5 @@
+#include "webcface/common/internal/message/func.h"
 #include "webcface/internal/func_internal.h"
-#include "webcface/common/internal/message.h"
 
 WEBCFACE_NS_BEGIN
