+++ client/src/log.cc
@@ -1,7 +1,8 @@
+#include "webcface/common/internal/message/pack.h"
 #include "webcface/log.h"
+#include "webcface/common/internal/message/log.h"
 #include "webcface/internal/logger.h"
 #include "webcface/internal/client_internal.h"
-#include "webcface/common/internal/message.h"
 
 WEBCFACE_NS_BEGIN
