+++ server-store/src/websock.cc
@@ -8,7 +8,6 @@
 #else
 #include "webcface/common/webcface-config.h"
 #endif
-#include "webcface/common/internal/message.h"
 #include "webcface/internal/unix_path.h"
 #include <memory>
 #include <thread>
